+++ R/utils-predict.R
@@ -34,8 +34,8 @@
 #' @param x `double` predicted survival
 #' @param n `integer(1)` number of patients/observation per interval
 #' @return `double`, cutpoints
-#' @examples
 #' @noRd
+#' @examples
 #' x <- seq(0, 1, length.out = 10)
 #' .cutpoints(x, n = 2)
 .cutpoints <- function(x, n = 50L) {